docs: add a TODO.md
This will be pivotal to help me capture thoughts about "what needs to be done"
and then make sure they don't float away while I'm hacking away on
something else. Could use a site like jira or whatnot
but in keeping with the self-contained monorepo
I think its fine that this docuement
lives with the codebase.