feat: add turbo repo, make `turbo build` task
The turbo build task will run `npm run build` in each of the child
workspaces of this npm monorepo. This the of feature unlocks the
security of knowing our commits always build. I find that this
type of mechanic encourages smaller commits and incremental
improvement. In the following commits we'll introduce
other turbo tasks and even leverage these tasks
in git hooks so that they are regularly part
of our workflow and don't fall vitim to
becomeing an afterthought of process.