feat: create code quality utilities
This commit includes a preinstall script and postinstall script that
help maintain code quality in two ways. The PREinstall script
ensures the author new to the team is using the the same
node runtime as required by the team. The second script
will migrate some git hooks into the local `.git`
folder (which isn't tracked by default).
These hooks ensure that the code can
be built before it's checked into
version control or pushed to
a shared remote.